+++ source/player.cpp
@@ -7,7 +7,7 @@
 extern int posX, posY;
 extern int pressureCount;
 extern int lastMoveDir;
-extern long lastInputTime;
+extern Uint32 lastInputTime;
 extern int frame;
 extern bool playerPlaced;
 extern bool displayName;
@@ -27,7 +27,7 @@ void cleanup();
 void changeTextToDead();
 class Level* getCurrentLevel();
 bool requestMove(int x, int y, int xChange, int yChange, Object* obj);
-long getTicks();
+Uint32 getTicks();
 extern double delta;
 Object* objectInit(unsigned int id, int x, int y, int moveDir, int moveFraction);
 //The player
